landing page and folder restructure
