+++ app/app-root.xml
@@ -1,2 +1,2 @@
-<Frame defaultPage="main-page">
+<Frame defaultPage="/screens/landing-page/landing-page">
 </Frame>
