+++ app/screens/landing-page/landing-page.ts
@@ -1,5 +1,5 @@
 import { EventData, Page } from '@nativescript/core'
-import { HelloWorldModel } from './main-view-model'
+import { HelloWorldModel } from './landing-page-model'
 
 export function navigatingTo(args: EventData) {
   const page = <Page>args.object
