+++ app/screens/landing-page/landing-page.xml
@@ -1,5 +1,5 @@
 <Page xmlns="http://schemas.nativescript.org/tns.xsd" navigatingTo="navigatingTo">
-    <ActionBar title="My App" icon="" />
+    <ActionBar title="Landing Page" icon="" />
 
     <StackLayout class="p-20">
         <Label text="Tap the button" class="h1 text-center" />
