--- app/splash-page.xml
@@ -1,5 +1,0 @@
-<Page xmlns="http://schemas.nativescript.org/tns.xsd" class="page">
-  <StackLayout>
-    <Label text="Loading..." class="h1 text-center" />
-  </StackLayout>
-</Page>
